+++ tests/framework/components/test_manager.py
@@ -2,7 +2,7 @@ from typing import Any, Dict
 
 import pytest
 
-from tests.helper_components import MockComponentA, MockComponentB, MockGenericComponent
+from tests.helpers import MockComponentA, MockComponentB, MockGenericComponent
 from vivarium.framework.components.manager import (
     ComponentConfigError,
     ComponentManager,
