+++ tests/framework/test_time.py
@@ -5,12 +5,7 @@ import numpy as np
 import pandas as pd
 import pytest
 
-from tests.helper_components import (
-    Listener,
-    MockComponentA,
-    MockComponentB,
-    MockGenericComponent,
-)
+from tests.helpers import Listener, MockComponentA, MockComponentB, MockGenericComponent
 from vivarium.framework.engine import SimulationContext as SimulationContext_
 from vivarium.framework.event import Event
 from vivarium.framework.time import SimulationClock, get_time_stamp
